+++ demos/friction.js
@@ -18,7 +18,7 @@ export function friction(manager) {
   manager.create({
     "transform": new Transform(200, 300, degToRad(20)),
     "movable": new Movable(),
-    "bounds": new BoundingBox(),
+    "bound": new BoundingBox(),
     "body": body1,
     "sprite": new Sprite(
       new BoxGeometry(300, 20),
@@ -29,7 +29,7 @@ export function friction(manager) {
   manager.create({
     "transform": new Transform(200, 200, degToRad(20)),
     "movable": new Movable(),
-    "bounds": new BoundingBox(),
+    "bound": new BoundingBox(),
     "body": body2,
     "sprite": new Sprite(
       new BoxGeometry(300, 20),
@@ -39,7 +39,7 @@ export function friction(manager) {
   manager.create({
     "transform": new Transform(100, 100, degToRad(20)),
     "movable": new Movable(),
-    "bounds": new BoundingBox(),
+    "bound": new BoundingBox(),
     "body": new Box(50, 20),
     "sprite": new Sprite(
       new BoxGeometry(50, 20),
@@ -49,7 +49,7 @@ export function friction(manager) {
   manager.create({
     "transform": new Transform(100, 237),
     "movable": new Movable(),
-    "bounds": new BoundingBox(),
+    "bound": new BoundingBox(),
     "body": new Ball(15),
     "sprite": new Sprite(
       new CircleGeometry(15),
@@ -57,8 +57,8 @@ export function friction(manager) {
     )
   })
 
-  body1.type = Body2D.STATIC
-  body2.type = Body2D.STATIC
+  Box.setType(body1, Box.STATIC)
+  Box.setType(body2, Box.STATIC)
 
-  manager.getResource("world").gravity = 980
+  manager.getResource("gravity").y = 900
 }
