+++ demos/index.html
@@ -22,7 +22,7 @@
       background: black
     }
   </style>
-  <script src="./main.js" type="module"></script>
+  <script src="./index.js" type="module"></script>
   <script type="sourcemap">
     {
         "imports":{
