+++ demos/restitution.js
@@ -7,29 +7,29 @@ import {
   BoxGeometry,
   BasicMaterial
 } from "/src/index.js"
-import {makePlatform} from "./utils.js"
+import { makePlatform } from "./utils.js"
 
 export function restitution(manager) {
   const rect = manager.getResource("renderer")
-  
+
   let body1 = new Box(50, 50)
   let body2 = new Box(50, 50)
   let body3 = new Box(50, 50)
   let body4 = new Box(50, 50)
   let body5 = new Box(50, 50)
   let body6 = new Box(50, 50)
-  
+
   body1.restitution = 1
   body2.restitution = 0.8
   body3.restitution = 0.6
   body4.restitution = 0.4
   body5.restitution = 0.2
   body6.restitution = 0
-  
+
   manager.create({
     "transform": new Transform(100, 300),
     "movable": new Movable(),
-    "bounds": new BoundingBox(),
+    "bound": new BoundingBox(),
     "body": body1,
     "sprite": new Sprite(
       new BoxGeometry(50, 50),
@@ -39,7 +39,7 @@ export function restitution(manager) {
   manager.create({
     "transform": new Transform(200, 300),
     "movable": new Movable(),
-    "bounds": new BoundingBox(),
+    "bound": new BoundingBox(),
     "body": body2,
     "sprite": new Sprite(
       new BoxGeometry(50, 50),
@@ -47,9 +47,9 @@ export function restitution(manager) {
     )
   })
   manager.create({
-    "transform": new Transform(300,300),
+    "transform": new Transform(300, 300),
     "movable": new Movable(),
-    "bounds": new BoundingBox(),
+    "bound": new BoundingBox(),
     "body": body3,
     "sprite": new Sprite(
       new BoxGeometry(50, 50),
@@ -57,9 +57,9 @@ export function restitution(manager) {
     )
   })
   manager.create({
-    "transform": new Transform(400,300),
+    "transform": new Transform(400, 300),
     "movable": new Movable(),
-    "bounds": new BoundingBox(),
+    "bound": new BoundingBox(),
     "body": body4,
     "sprite": new Sprite(
       new BoxGeometry(50, 50),
@@ -69,7 +69,7 @@ export function restitution(manager) {
   manager.create({
     "transform": new Transform(500, 300),
     "movable": new Movable(),
-    "bounds": new BoundingBox(),
+    "bound": new BoundingBox(),
     "body": body5,
     "sprite": new Sprite(
       new BoxGeometry(50, 50),
@@ -79,7 +79,7 @@ export function restitution(manager) {
   manager.create({
     "transform": new Transform(600, 300),
     "movable": new Movable(),
-    "bounds": new BoundingBox(),
+    "bound": new BoundingBox(),
     "body": body6,
     "sprite": new Sprite(
       new BoxGeometry(50, 50),
@@ -93,5 +93,5 @@ export function restitution(manager) {
     rect.width,
     50
   )
-  manager.getResource("world").gravity = 980
+  manager.getResource("gravity").y = 900
 }
