+++ demos/stacking.js
@@ -24,8 +24,7 @@ export function stacking(manager) {
   stack(800, 100, 50, 50, 6, 5, manager)
   stack(1000, 100, 50, 50, 4, 5, manager)
   
-  manager.getResource("world").gravity = 980
-
+  manager.getResource("gravity").y = 900
 }
 
 function stack(x, y, w, h, no, spacing, manager) {
@@ -33,7 +32,7 @@ function stack(x, y, w, h, no, spacing, manager) {
     manager.create({
       "transform": new Transform(x, y + (h + spacing) * i),
       "movable": new Movable(),
-      "bounds": new BoundingBox(),
+      "bound": new BoundingBox(),
       "body": new Box(w, h),
       "sprite": new Sprite(
         new BoxGeometry(w, h),
