+++ demos/triangle.js
@@ -11,12 +11,12 @@ import { makePlatform } from "./utils.js"
 
 export function triangle(manager) {
   const rect = manager.getResource("renderer")
-  const angle = Math.PI/2
+  const angle = Math.PI / 2
   const base = 200
   manager.create({
     "transform": new Transform(rect.width / 2, 300),
     "movable": new Movable(),
-    "bounds": new BoundingBox(),
+    "bound": new BoundingBox(),
     "body": new Trigon(base, (base / 2) * Math.sin(angle), angle),
     "sprite": new Sprite(
       new TriangleGeometry(base, (base / 2) * Math.sin(angle), angle),
@@ -30,5 +30,5 @@ export function triangle(manager) {
     rect.width,
     50
   )
-  manager.getResource("world").gravity = 900
+  manager.getResource("gravity").y = 900
 }
